requirements.txt
flask
scikit-learn
gunicorn
pandas
sqlalchemy
psycopg2-binary
cohere
python-dotenv